Test-EwsEndpoint.ps1
This script tests the connectivity of an EWS endpoint(s) configured in MigrationWiz project connector(s). Using the unique project information an EWS endpoint URL is generated based on the OWA URL of a connector. The EWS endpoint URL is then used to confirm the EWS endpoint exists, and can be connected to using the information in the project.